Moved interfaces and typeguards into correct locations
