+++ src/interfaces/typeGuards.ts
@@ -22,6 +22,7 @@ import {
     TicketPeriod,
     MultiOperatorInfo,
     MultiOperatorInfoWithErrors,
+    TermTimeAttribute,
 } from './index';
 
 import { FareType, FareTypeWithErrors } from '../pages/api/fareType';
@@ -31,6 +32,7 @@ import { FareStagesAttribute, FareStagesAttributeWithErrors } from '../pages/api
 import { InputCheck } from '../pages/stageNames';
 import { TicketPeriodWithErrors } from '../pages/api/productDateInformation';
 import { MultipleOperatorsAttribute, MultipleOperatorsAttributeWithErrors } from '../pages/api/searchOperators';
+import { TermTimeAttributeWithErrors } from '../pages/termTime';
 
 export const isNotEmpty = <T>(value: T | null | undefined): value is T => value !== null && value !== undefined;
 
@@ -182,3 +184,8 @@ export const isMultiOperatorInfoWithErrors = (
     multiOperatorInfo: MultiOperatorInfo[] | MultiOperatorInfoWithErrors | undefined,
 ): multiOperatorInfo is MultiOperatorInfoWithErrors =>
     multiOperatorInfo !== undefined && (multiOperatorInfo as MultiOperatorInfoWithErrors).errors !== undefined;
+
+export const isTermTimeAttributeWithErrors = (
+    termTime: undefined | TermTimeAttribute | TermTimeAttributeWithErrors,
+): termTime is TermTimeAttributeWithErrors =>
+    !!termTime && (termTime as TermTimeAttributeWithErrors).errors !== undefined;
